+++ src/Translation/Term/Metabox/SourceTermSaveContext.php
@@ -97,7 +97,7 @@ final class SourceTermSaveContext implements \ArrayAccess {
 			self::RELATED_BLOGS    => [],
 		];
 
-		$source_site_id = (int) get_current_blog_id();
+		$source_site_id = get_current_blog_id();
 
 		$related_blogs = $this->site_relations->get_related_site_ids( $source_site_id );
 		if ( empty( $related_blogs ) ) {
